Created components and added top bar for home page
