--- src/components/Cart/index.js
@@ -1,3 +1,0 @@
-import react from "react";
-
-export default function Home() {}
