--- src/components/ProductDetail/index.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-export default function Home() {
-
-return{
-
-
-
-
-
-
-    }
-
-}
