+++ src/components/home/Home.css
@@ -1,0 +1,17 @@
+.Home-header-bar {
+    padding: 10px 0;
+    background: rgba(155, 16, 16, 0.7) none repeat scroll 0 0;
+    margin-top: 10px;
+    z-index: 1;
+}
+.Home-header-container {
+    padding-right: 15px;
+    padding-left: 15px;
+    margin-right: auto;
+    margin-left: auto;
+}
+
+.Home-header-text{
+    padding-top: 10px;
+    padding-left: 800px;
+}
